# backups/20250827T074527Z/openapi/openapi_20250827T074527Z.json
{"openapi":"3.1.0","info":{"title":"PortPulse API","version":"0.1.1"},"paths":{"/v1/health":{"get":{"tags":["meta"],"summary":"Liveness/Readiness","operationId":"health_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/hs/{code}/imports":{"get":{"tags":["hs","hs"],"summary":"HS imports (demo)","operationId":"hs_imports_v1_hs__code__imports_get","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}},{"name":"frm","in":"query","required":true,"schema":{"type":"string","minLength":2,"maxLength":3,"description":"Origin ISO-2/3","title":"Frm"},"description":"Origin ISO-2/3"},{"name":"to","in":"query","required":true,"schema":{"type":"string","minLength":2,"maxLength":3,"description":"Destination ISO-2/3","title":"To"},"description":"Destination ISO-2/3"},{"name":"months","in":"query","required":false,"schema":{"type":"integer","maximum":36,"minimum":1,"default":6,"title":"Months"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","pattern":"^(json|csv)$","default":"json","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"head":{"tags":["hs","hs"],"summary":"Hs Imports Head","operationId":"hs_imports_head_v1_hs__code__imports_head","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}},{"name":"frm","in":"query","required":true,"schema":{"type":"string","title":"Frm"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","title":"To"}},{"name":"months","in":"query","required":false,"schema":{"type":"integer","default":6,"title":"Months"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","default":"json","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/alerts":{"get":{"tags":["alerts","alerts"],"summary":"Dwell change alerts (v1)","operationId":"get_alerts_v1_ports__unlocode__alerts_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"window","in":"query","required":false,"schema":{"type":"integer","maximum":60,"minimum":7,"default":14,"title":"Window"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/overview":{"get":{"tags":["ports"],"summary":"Get Overview Csv","operationId":"get_overview_csv_v1_ports__unlocode__overview_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","default":"csv","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortOverview"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"head":{"tags":["ports"],"summary":"Head Overview Csv","operationId":"head_overview_csv_v1_ports__unlocode__overview_head","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","default":"csv","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/calls":{"get":{"tags":["ports","ports"],"summary":"Port Calls","description":"Get detailed port calls information with pagination support","operationId":"port_calls_v1_ports__unlocode__calls_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PortCallExpanded"},"title":"Response Port Calls V1 Ports  Unlocode  Calls Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/calls/processed":{"get":{"tags":["ports","ports"],"summary":"Processed Port Calls","description":"Get processed port calls information with additional calculated fields and pagination support","operationId":"processed_port_calls_v1_ports__unlocode__calls_processed_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PortCallProcessed"},"title":"Response Processed Port Calls V1 Ports  Unlocode  Calls Processed Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/trend":{"get":{"tags":["ports","ports"],"summary":"Daily trend (JSON/CSV)","operationId":"trend_v1_ports__unlocode__trend_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}},{"name":"fields","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"csv/json fields, comma-separated","title":"Fields"},"description":"csv/json fields, comma-separated"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","pattern":"^(json|csv)$","default":"json","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/dwell":{"get":{"tags":["ports","ports"],"summary":"Daily dwell hours","operationId":"dwell_v1_ports__unlocode__dwell_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ports/{unlocode}/snapshot":{"get":{"tags":["ports","ports"],"summary":"Latest snapshot (top-level not null)","operationId":"snapshot_v1_ports__unlocode__snapshot_get","parameters":[{"name":"unlocode","in":"path","required":true,"schema":{"type":"string","title":"Unlocode"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AlertItem":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"metric":{"type":"string","title":"Metric","default":"dwell_hours"},"delta":{"type":"number","title":"Delta"},"severity":{"type":"string","title":"Severity","default":"info"},"explain":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explain"}},"type":"object","required":["date","delta"],"title":"AlertItem"},"AlertsResponse":{"properties":{"unlocode":{"type":"string","title":"Unlocode"},"window_days":{"type":"integer","title":"Window Days","default":14},"items":{"items":{"$ref":"#/components/schemas/AlertItem"},"type":"array","title":"Items","default":[]}},"type":"object","required":["unlocode"],"title":"AlertsResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PortCallExpanded":{"properties":{"call_id":{"type":"string","title":"Call Id","description":"靠港记录 ID"},"unlocode":{"type":"string","title":"Unlocode","description":"港口 UN/LOCODE"},"vessel_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vessel Name","description":"船名"},"imo":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Imo","description":"IMO 号"},"mmsi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mmsi","description":"MMSI"},"status":{"anyOf":[{"type":"string","enum":["expected","arrived","anchorage","berthed","sailed"]},{"type":"null"}],"title":"Status"},"eta":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Eta"},"etd":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Etd"},"ata":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Ata"},"atb":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Atb"},"atd":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Atd"},"berth":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Berth"},"terminal":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Terminal"},"last_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Updated At"}},"type":"object","required":["call_id","unlocode"],"title":"PortCallExpanded","description":"单船靠港过程的扩展信息（明细视图）"},"PortCallProcessed":{"properties":{"call_id":{"type":"string","title":"Call Id"},"unlocode":{"type":"string","title":"Unlocode"},"phase":{"anyOf":[{"type":"string","enum":["waiting","berthing","turnaround"]},{"type":"null"}],"title":"Phase"},"wait_hours":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Wait Hours"},"service_time_hours":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Service Time Hours","description":"作业时长（小时）"},"turnaround_hours":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Turnaround Hours","description":"到港到离港总周期（小时）"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["call_id","unlocode"],"title":"PortCallProcessed","description":"经过加工处理后的指标（便于分析/看板）"},"PortOverview":{"properties":{"unlocode":{"type":"string","title":"Unlocode","description":"联合国港口代码，5 位"},"port_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Port Name","description":"港口名称"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country","description":"国家/地区"},"arrivals_7d":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Arrivals 7D","description":"近 7 日到港数"},"departures_7d":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Departures 7D","description":"近 7 日离港数"},"waiting_vessels":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Waiting Vessels","description":"当前锚地等待船舶数"},"avg_wait_hours":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Avg Wait Hours","description":"平均等待时长（小时）"},"avg_berth_hours":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Avg Berth Hours","description":"平均靠泊时长（小时）"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At","description":"数据快照时间戳"}},"type":"object","required":["unlocode"],"title":"PortOverview","description":"端口概览（用于 /v1/ports/{unlocode}/overview 的 JSON 结构）"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}